Seperated Base Manipalanet Data Provider into 2 providers for easier use.
