+++ src/eXpansion/Framework/GameManiaplanet/Resources/config/data_providers.yml
@@ -1,18 +1,27 @@
 parameters:
-    expansion.framework.game_maniaplanet.data_providers.base.class: eXpansion\Framework\GameManiaplanet\DataProviders\BaseDataProvider
-    expansion.framework.game_maniaplanet.data_providers.base.interface: eXpansion\Framework\GameManiaplanet\DataProviders\Listener\BaseDataListenerInterface
+    expansion.framework.game_maniaplanet.data_providers.map.class: eXpansion\Framework\GameManiaplanet\DataProviders\MapDataProvider
+    expansion.framework.game_maniaplanet.data_providers.map.interface: eXpansion\Framework\GameManiaplanet\DataProviders\Listener\MapDataListenerInterface
+
+    expansion.framework.game_maniaplanet.data_providers.match.class: eXpansion\Framework\GameManiaplanet\DataProviders\MatchDataProvider
+    expansion.framework.game_maniaplanet.data_providers.match.interface: eXpansion\Framework\GameManiaplanet\DataProviders\Listener\MatchDataListenerInterface
 
 services:
-    expansion.framework.game_maniaplanet.data_providers.base:
-        class: "%expansion.framework.game_maniaplanet.data_providers.base.class%"
+    expansion.framework.game_maniaplanet.data_providers.map:
+        class: "%expansion.framework.game_maniaplanet.data_providers.map.class%"
         arguments:
             - "@expansion.framework.core.storage.map"
         tags:
-            - {name: expansion.data_provider, provider: "expansion.mp.base", interface: "%expansion.framework.game_maniaplanet.data_providers.base.interface%"}
+            - {name: expansion.data_provider, provider: "expansion.mp.map", interface: "%expansion.framework.game_maniaplanet.data_providers.map.interface%"}
             - {name: expansion.data_provider.compatibility, title: ALL, gamemode: script}
             - {name: expansion.data_provider.listener, event_name: Maniaplanet.StartMap_Start, method: onStartMapStart}
             - {name: expansion.data_provider.listener, event_name: Maniaplanet.StartMap_End, method: onStartMapEnd}
             - {name: expansion.data_provider.listener, event_name: Maniaplanet.EndMap_Start, method: onEndMapStart}
             - {name: expansion.data_provider.listener, event_name: Maniaplanet.EndMap_End, method: onEndMapEnd}
+
+    expansion.framework.game_maniaplanet.data_providers.match:
+        class: "%expansion.framework.game_maniaplanet.data_providers.match.class%"
+        tags:
+            - {name: expansion.data_provider, provider: "expansion.mp.match", interface: "%expansion.framework.game_maniaplanet.data_providers.match.interface%"}
+            - {name: expansion.data_provider.compatibility, title: ALL, gamemode: script}
             - {name: expansion.data_provider.listener, event_name: Maniaplanet.StartMatch_Start, method: onStartMatchStart}
             - {name: expansion.data_provider.listener, event_name: Maniaplanet.StartMatch_End, method: onStartMatchEnd}
